prepopulate create form data

+++ biblehelp/static/scripts.js
@@ -1,37 +1,48 @@
 const csrftoken = getCookie('csrftoken');
 const goalContainer = document.querySelector('.goal-container');
+const setGoalBtns = document.querySelectorAll('.set-goal');
 
-goalContainer.addEventListener('click', (e) => {
-    // See if delete button was clicked
-    if (e.target.classList.contains('delete')) {
-        const goalId = e.target.parentNode.querySelector('input').value;
-        const itemContainer = e.target.parentNode.parentNode;
+// Check to see if setGoalBtns is on the DOM
+if (setGoalBtns) {
+    setGoalBtns.forEach(btn => btn.addEventListener('click', (e) => {
+        console.log(e.target);
+    }));
+}
 
-        // Send fetch request to delete goal
-        fetch('/goals', {
-            method: 'DELETE',
-            credentials: 'include',
-            headers: { 'X-CSRFToken': csrftoken},
-            body: JSON.stringify({
-                goalId: goalId
+// Check to see if goalContainer is on the DOM
+if (goalContainer) {
+    goalContainer.addEventListener('click', (e) => {
+        // See if delete button was clicked
+        if (e.target.classList.contains('delete')) {
+            const goalId = e.target.parentNode.querySelector('input').value;
+            const itemContainer = e.target.parentNode.parentNode;
+
+            // Send fetch request to delete goal
+            fetch('/goals', {
+                method: 'DELETE',
+                credentials: 'include',
+                headers: { 'X-CSRFToken': csrftoken },
+                body: JSON.stringify({
+                    goalId: goalId
+                })
             })
-        })
-        .then(response => {
-            if (response.ok) {
-                return response.json();
-            } else {
-                return Promise.reject(response.statusText);
-            }
-        })
-        .then(data => {
-            if (data.message === 'success') {
-                itemContainer.remove();
-            }
-        })
-        // on error would return 'The error is:' + status text
-        .catch(error => console.log('The error is:', error));
-    };
-})
+                .then(response => {
+                    if (response.ok) {
+                        return response.json();
+                    } else {
+                        return Promise.reject(response.statusText);
+                    }
+                })
+                .then(data => {
+                    if (data.message === 'success') {
+                        itemContainer.remove();
+                    }
+                })
+                // on error would return 'The error is:' + status text
+                .catch(error => console.log('The error is:', error));
+        };
+    })
+}
